# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CreviaTube API
  description: |
    Advanced content creator platform API for monetizing content and managing campaigns.
    
    ## Authentication
    Most endpoints require authentication. Include your session cookie or use the login endpoint to authenticate.
    
    ## Rate Limiting
    - Authentication endpoints: 5 attempts per 15 minutes
    - API endpoints: 100 requests per 15 minutes
    - Sensitive operations: 10 requests per hour
    - Tracking events: 30 events per minute
    - Admin operations: 50 requests per 15 minutes
    
    ## Error Codes
    - `400` - Bad Request
    - `401` - Unauthorized
    - `403` - Forbidden
    - `404` - Not Found
    - `429` - Too Many Requests
    - `500` - Internal Server Error
  version: 1.0.0
  contact:
    name: CreviaTube Support
    email: support@creviatube.com
    url: https://creviatube.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.creviatube.com
    description: Production server
security:
  - sessionAuth: []
paths:
...
  /api/analytics/stats:
    get:
      tags:
        - Analytics
      summary: Get analytics statistics
      description: Retrieve overall analytics statistics
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Analytics statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsStats'
...
  /api/visualization/charts:
    get:
      tags:
        - Visualization
      summary: Get all charts
      description: Retrieve all available charts
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Charts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  charts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Chart'
  /api/visualization/reports:
    get:
      tags:
        - Visualization
      summary: Get all reports
      description: Retrieve all custom reports
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Reports retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  reports:
                    type: array
                    items:
                      $ref: '#/components/schemas/Report'

  # ===== PERFORMANCE ENDPOINTS =====
...
  /api/cache/stats:
    get:
      tags:
        - Cache
      summary: Get cache statistics
      description: Retrieve Redis cache statistics
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Cache statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheStats'
  /api/cache/clear:
    post:
      tags:
        - Cache
      summary: Clear cache
      description: Clear all cached data
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Cache cleared successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Cache cleared successfully"
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
...